refactor: extract ACP runtime resume state
